requirements.txt
streamlit
torch
torchvision
numpy
pandas
scikit-learn
matplotlib
tqdm
pillow
imagehash
